projects/Genesis/pyproject.toml
[build-system]
requires = ["setuptools==80.9.0", "wheel==0.45.1"]
build-backend = "setuptools.build_meta"

[project]
name = "genesis-evolve"
version = "0.1.0"
description = "Genesis: LLM-Driven Program Evolution Platform"
authors = [
    {name = "George Pearse", email = "george@genesis.ai"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = "==3.12.*"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "openai==1.58.1",
    "numpy==2.2.6",
    "pandas==2.2.3",
    "anthropic==0.42.0",
    "requests==2.32.5",
    "boto3==1.39.4",
    "pydantic==2.10.1",
    "backoff==2.2.1",
    "python-dotenv==1.0.1",
    "instructor==1.7.2",
    "python-Levenshtein>=0.20.0",
    "radon==6.0.1",
    "unidiff==0.7.5",
    "dill==0.4.0",
    "hydra-core==1.3.2",
    "matplotlib==3.9.2",
    "networkx==3.5",
    "seaborn==0.13.2",
    "moviepy==2.1.1",
    "scikit-learn==1.7.1",
    "adjustText>=1.0.0",
    "markdown==3.8.2",
    "aiofiles==24.1.0",
    "e2b>=1.0.0",
    "e2b-code-interpreter>=0.0.11",
    "psycopg[binary]==3.2.3",
    "google-generativeai",
    "mcp>=1.0.0",
    "clickhouse-connect>=0.7.0",
]

[project.scripts]
genesis = "genesis.cli:main"

[tool.setuptools]
script-files = ["genesis/genesis_launch", "genesis/genesis_mcp"]

[tool.setuptools.packages.find]
include = ["genesis", "genesis.*"]

[tool.setuptools.package-data]
"*" = ["*"]

[tool.uv]
dev-dependencies = [
    "pytest==8.3.3",
    "ruff==0.8.0",
    "zuban==0.2.3",
    "pre-commit==4.0.1",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E203", "E501"]

[tool.zuban]
target-version = "py3.12"
include = ["genesis", "examples"]
exclude = ["results", "tests"]
